requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2023.10.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
pyfiglet==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
termcolor==2.4.0
torch==2.1.1
torchvision==0.16.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
